Rework SubARS for DD work
Keeping the subtype property in the type of SubARS causes some issues,
for instance when we would like to create a set of SubARSes (e.g.
ARS.components). Additionally, the attempt of using union_rel to handle
a SubARS with a different index type doesn't work very well; the more
reasonable definition simply requires the index type to be the same,
and the restriction and closure requirements to range over each individual
rewrite relation.